requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
passlib[bcrypt]
python-jose[cryptography]
python-dotenv
keyring
google-cloud-secret-manager
pytest-asyncio
twilio
